+++ items/MenuController.php
@@ -64,7 +64,7 @@ class MenuController extends Controller
         $model = new Menu;
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->menu_id]);
+            return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('create', [
                     'model' => $model,
@@ -82,10 +82,10 @@ class MenuController extends Controller
     {
         $model = $this->findModel($id);
         if ($model->menuParent) {
-            $model->menu_parent_name = $model->menuParent->menu_name;
+            $model->parent_name = $model->menuParent->name;
         }
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->menu_id]);
+            return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('update', [
                     'model' => $model,