+++ migration/m140602_111327_create_menu_table.php
@@ -7,11 +7,12 @@ class m140602_111327_create_menu_table extends \yii\db\Migration
     public function safeUp()
     {
         $this->createTable('{{%menu}}', [
-            'menu_id'=>  Schema::TYPE_PK,
-            'menu_name'=>  Schema::TYPE_STRING.'(128) NOT NULL',
-            'menu_parent'=>  Schema::TYPE_INTEGER,
-            'menu_route'=>  Schema::TYPE_STRING.'(256)',
-            'FOREIGN KEY (menu_parent) REFERENCES {{%menu}}(menu_id) ON DELETE SET NULL ON UPDATE CASCADE',
+            'id'=>  Schema::TYPE_PK,
+            'name'=>  Schema::TYPE_STRING.'(128) NOT NULL',
+            'parent'=>  Schema::TYPE_INTEGER,
+            'route'=>  Schema::TYPE_STRING.'(256)',
+            'data'=>  Schema::TYPE_TEXT,
+            'FOREIGN KEY (parent) REFERENCES {{%menu}}(id) ON DELETE SET NULL ON UPDATE CASCADE',
         ]);
     }