+++ models/Menu.php
@@ -3,21 +3,22 @@
 namespace mdm\admin\models;
 
 use Yii;
+use mdm\admin\components\AccessHelper;
 
 /**
  * This is the model class for table "menu".
  *
- * @property integer $menu_id
- * @property string $menu_name
- * @property integer $menu_parent
- * @property string $menu_route
+ * @property integer $id
+ * @property string $name
+ * @property integer $parent
+ * @property string $route
  *
  * @property Menu $menuParent
  * @property Menu[] $menus
  */
 class Menu extends \yii\db\ActiveRecord
 {
-    public $menu_parent_name;
+    public $parent_name;
 
     /**
      * @inheritdoc
@@ -33,32 +34,33 @@ class Menu extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            [['menu_name'], 'required'],
-            [['menu_parent_name'], 'filterParent'],
-            [['menu_parent_name'], 'in', 
-                'range' => self::find()->select(['menu_name'])->column(),
+            [['name'], 'required'],
+            [['parent_name'], 'filterParent'],
+            [['parent_name'], 'in',
+                'range' => self::find()->select(['name'])->column(),
                 'message' => 'Menu "{value}" not found.'],
-            [['menu_parent'], 'default'],
-            [['menu_name'], 'string', 'max' => 128],
-            [['menu_route'], 'string', 'max' => 256]
+            [['parent', 'route', 'data'], 'default'],
+            [['route'], 'in',
+                'range' => AccessHelper::getSavedRoutes(),
+                'message' => 'Route "{value}" not found.']
         ];
     }
 
     public function filterParent()
     {
-        $value = $this->menu_parent_name;
-        $parent = self::findOne(['menu_name' => $value]);
+        $value = $this->parent_name;
+        $parent = self::findOne(['name' => $value]);
         if ($parent) {
-            $id = $this->menu_id;
-            $parent_id = $parent->menu_id;
+            $id = $this->id;
+            $parent_id = $parent->id;
             while ($parent) {
-                if ($parent->menu_id == $id) {
-                    $this->addError('menu_parent_name', 'Loop detected.');
+                if ($parent->id == $id) {
+                    $this->addError('parent_name', 'Loop detected.');
                     return;
                 }
                 $parent = $parent->menuParent;
             }
-            $this->menu_parent = $parent_id;
+            $this->parent = $parent_id;
         }
     }
 
@@ -68,11 +70,11 @@ class Menu extends \yii\db\ActiveRecord
     public function attributeLabels()
     {
         return [
-            'menu_id' => 'Menu ID',
-            'menu_name' => 'Menu Name',
-            'menu_parent' => 'Menu Parent',
-            'menu_parent_name' => 'Menu Parent',
-            'menu_route' => 'Menu Route',
+            'id' => 'ID',
+            'name' => 'Name',
+            'parent' => 'Parent',
+            'parent_name' => 'Parent',
+            'route' => 'Route',
         ];
     }
 
@@ -81,7 +83,7 @@ class Menu extends \yii\db\ActiveRecord
      */
     public function getMenuParent()
     {
-        return $this->hasOne(Menu::className(), ['menu_id' => 'menu_parent']);
+        return $this->hasOne(Menu::className(), ['id' => 'parent']);
     }
 
     /**
@@ -89,6 +91,6 @@ class Menu extends \yii\db\ActiveRecord
      */
     public function getMenus()
     {
-        return $this->hasMany(Menu::className(), ['menu_parent' => 'menu_id']);
+        return $this->hasMany(Menu::className(), ['parent' => 'id']);
     }
 }