+++ views/menu/_search.php
@@ -17,14 +17,16 @@ use yii\widgets\ActiveForm;
         'method' => 'get',
     ]); ?>
 
-    <?= $form->field($model, 'menu_id') ?>
+    <?= $form->field($model, 'id') ?>
 
-    <?= $form->field($model, 'menu_name') ?>
+    <?= $form->field($model, 'name') ?>
 
-    <?= $form->field($model, 'menu_parent') ?>
+    <?= $form->field($model, 'parent') ?>
 
-    <?= $form->field($model, 'menu_route') ?>
+    <?= $form->field($model, 'route') ?>
 
+    <?= $form->field($model, 'data') ?>
+    
     <div class="form-group">
         <?= Html::submitButton('Search', ['class' => 'btn btn-primary']) ?>
         <?= Html::resetButton('Reset', ['class' => 'btn btn-default']) ?>